% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{authorpage}
\alias{authorpage}
\title{Page with authors}
\format{A list format webpage with author names}
\source{
EPSA Paper authors. Scraped from EPSA conference program.
}
\usage{
authorpage
}
\description{
Page with authors
}
\examples{
\dontrun{
 authorpage
}
}
\keyword{datasets}
