% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{epsa_paper}
\alias{epsa_paper}
\title{EPSA Papers}
\format{A data frame with columns:
\describe{
 \item{title}{paper title}
 \item{section}{EPSA section}
 \item{abstracts}{Paper abstract}
 \item{panel}{Panel location}
 \item{url}{relative path to abstract}
}}
\source{
EPSA Papers. Scraped from EPSA conference program.
}
\usage{
epsa_paper
}
\description{
EPSA Papers
}
\examples{
\dontrun{
 epsa_paper
}
}
\keyword{datasets}
