% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{german_cabinets}
\alias{german_cabinets}
\title{German cabinets}
\format{
An object of class \code{list} of length 30.
}
\source{
\url{https://whogoverns.eu/cabinets/?view=pais&codigo=de}
}
\usage{
german_cabinets
}
\description{
A list of URLs to German cabinets listed on \url{https://whogoverns.eu}
}
\details{
A list. Names specifying cabinet names; values corresponding URLs.
}
\keyword{datasets}
