% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nyt_year_counts}
\alias{nyt_year_counts}
\title{Count of NYT articles containing the term "democracy" for years between 2000 and 2020}
\format{
An object of class \code{integer} of length 21.
}
\source{
\url{http://api.nytimes.com/svc/search/v2/articlesearch.json}
}
\usage{
nyt_year_counts
}
\description{
A vector of integer values named by years.
}
\details{
An integer vector. Names specifying year; values record article counts.
}
\keyword{datasets}
