% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{quotepage}
\alias{quotepage}
\title{Page with quotes}
\format{A list format webpage with quotes}
\source{
Quotes to scrape webpage
}
\usage{
quotepage
}
\description{
Page with quotes
}
\examples{
\dontrun{
 quotepage
}
}
\keyword{datasets}
