% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{titlepage}
\alias{titlepage}
\title{Page with titles}
\format{A list format webpage with paper titles}
\source{
EPSA Paper titles. Scraped from EPSA conference program.
}
\usage{
titlepage
}
\description{
Page with titles
}
\examples{
\dontrun{
 titlepage
}
}
\keyword{datasets}
