% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{epsa_paper}
\alias{epsa_paper}
\title{EPSA Papers}
\format{
A data frame with columns:
\describe{
 \item{title}{paper title}
 \item{author1}{first author}
 \item{author2}{second author}
 \item{author3}{third author}
 \item{author4}{fourth author}
 \item{author5}{fifth author}
 \item{author6}{sixth author}
 \item{institution1}{first author}
 \item{institution2}{second institution}
 \item{institution3}{third institution}
 \item{institution4}{fourth institution}
 \item{institution5}{fifth institution}
 \item{institution6}{sixth institution}
 \item{section}{EPSA section}
 \item{abstracts}{Paper abstract}
 \item{panel}{Panel location}
 \item{url}{relative path to abstract}
}
}
\source{
EPSA Papers. Scraped from EPSA conference program.
}
\usage{
epsa_paper
}
\description{
EPSA Papers
}
\examples{
\dontrun{
 epsa_paper
}
}
\keyword{datasets}
